Cohesion database views have been moved to their own new header (in the metrics parser).
