+++ plugins/cpp_metrics/model/CMakeLists.txt
@@ -6,6 +6,7 @@ message(WARNING "${cpp_PLUGIN_DIR}/model/include")
 
 set(ODB_SOURCES
   include/model/cppastnodemetrics.h
+  include/model/cppcohesionmetrics.h
   include/model/cppfilemetrics.h)
 
 generate_odb_files("${ODB_SOURCES}" "cpp")
